Efficiency update to code to prevent flickering
The distance sensor was taking too long to operate and messing with the display of the digits. Now it only senses every SENSE_COOLDOWN(=10) executions of the loop. This way we spend more time displaying the digits without the execution of the distance sensor in between and this will make the display brighter because it is on more of the time.